tmf: Ensure order of trace updated and trace range updated signals
When the trace index already exists and doesn't need to be build
from scratch the signal TmfTraceUpdatedSignal is sent once in the
method buildIndex() of the TmfCheckpointIndexer. However, buildIndex()
is called from the signal handler of class TmfTrace for the
signal TmfTraceRangeUpdatedSignal. That causes that signal
TmfTraceUpdatedSignal is sent before TmfTraceRangeUpdatedSignal
is processed by all other signal handlers. This causes problems
for example in the Histogram view where the trace end time is not
set correctly because this wrong order of signal handling.

This patch fixes that problem by wrapping the sending of signal
TmfTraceUpdatedSignal in a TmfEventRequest which is only dispatched
by the TmfEventProvider (=TmfTrace) after the first signal
(TmfTraceRangeUpdatedSignal) is processed by all signal handlers.

Change-Id: I96dfdb4cc678722ff5b48a8881c67f5a8c11ff36
Signed-off-by: Bernd Hufmann <Bernd.Hufmann@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/20216
Tested-by: Hudson CI
Reviewed-by: Genevieve Bastien <gbastien+lttng@versatic.net>
IP-Clean: Genevieve Bastien <gbastien+lttng@versatic.net>
Tested-by: Genevieve Bastien <gbastien+lttng@versatic.net>
Reviewed-by: Bernd Hufmann <bernd.hufmann@ericsson.com>