tmf: Split ITmfStateSystem.waitUntilBuilt() in separate methods
Decouple the return value of waitUntilBuilt() in a new isCancelled()
method, so that it can be checked independently.

This is because we are planning to add a new waitUntilBuilt(long) method
with a timeout, and that one will have to return a boolean to indicate
if it returned due to the thing being built, or due to the timeout.

We will still want to check for the cancellation state afterwards.
Furthermore, it would be very confusing if both methods (with and without
the timeout) returned a boolean that meant different things.

Change-Id: If7a954d2368d55c99e30e31e17e39b39bca5f432
Signed-off-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>
Reviewed-on: https://git.eclipse.org/r/20677
Reviewed-by: Patrick Tasse <patrick.tasse@gmail.com>
IP-Clean: Patrick Tasse <patrick.tasse@gmail.com>
Tested-by: Hudson CI