Systemtap: Warn when about to overrun a script.
Running a new instance of a running script terminates the
old one. If a user tries to run a script in this way,
a warning dialog now appears to ask the user if they
want to terminate the old run, rather than terminating
it automatically (See EBZ #394999).

Change-Id: I1f8798fe7ce964fa1e584e9cd285a213d69bed77
Signed-off-by: Andrew Ferrazzutti <aferrazz@redhat.com>
Reviewed-on: https://git.eclipse.org/r/20895
Tested-by: Hudson CI
Reviewed-by: Alexander Kurtakov <akurtako@redhat.com>
IP-Clean: Alexander Kurtakov <akurtako@redhat.com>
Tested-by: Alexander Kurtakov <akurtako@redhat.com>