tmf: Don't return -1 for interval end times in the TransientState
This might make sense locally to this class, but no other component
anywhere uses -1 for an end time value. Plus it breaks the contract
checked at construction that the end time should always be > the start
time.

Instead of returning a dummy -1, return the latest known end time. This
is just as "wrong", but eventually we could augment to API to indicate
if it read a real interval or a current state.

Change-Id: Ia9a23f990dce3301c721996e75219bc34ed1164f
Signed-off-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>
Reviewed-on: https://git.eclipse.org/r/20903
Reviewed-by: Patrick Tasse <patrick.tasse@gmail.com>
IP-Clean: Patrick Tasse <patrick.tasse@gmail.com>
Tested-by: Patrick Tasse <patrick.tasse@gmail.com>