tmf: Dispose analysis modules when the trace is disposed
There was a dispose() method in the StateSystemAnalysisModule, but
it came from HistoryBuilder, and was about disposing the state provider
only (which we do once the event request is complete, since we will not
need to insert more events in it). It had nothing to do with
ITmfComponent.dispose().

Add a "real" dipose method to the state system module, which disposes
the state system that was built, if any. And then also call that dispose
from TmfTrace.dispose().

This way the analysis module should:
- Correctly deregister from the TmfSignalManager, to be eligible for
  garbage collection.
- Dispose the statesystem object it contains, correctly closing any
  backend files.

Change-Id: I5836009f7172ab4ef9a41e22b5dbc0ef2fccbe4e
Signed-off-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>
Reviewed-on: https://git.eclipse.org/r/21072