TMF: Add method to get traces with build thread in Time graph view
This new method returns the traces for which a build thread will be started
in the abstract time graph view. This will allow a time graph view displaying
results from an experiment to override this method so that only one build
thread for the experiment starts.

Change-Id: I630424b3753f98924fc4c2dfc71a4902cddb0bee
Signed-off-by: Geneviève Bastien <gbastien+lttng@versatic.net>
Reviewed-on: https://git.eclipse.org/r/21887
Tested-by: Hudson CI
Reviewed-by: Marc-Andre Laperle <marc-andre.laperle@ericsson.com>
IP-Clean: Marc-Andre Laperle <marc-andre.laperle@ericsson.com>
Tested-by: Marc-Andre Laperle <marc-andre.laperle@ericsson.com>