Systemtap: Make some fixes to Probe & Function views.
-The icons next to Function view entries now correctly
represent the return type of functions and the variable
type of their parameters.

-Fix a bug that prevented some tapset functions, as well
as probe aliases, from being available in their appropriate
views (EBZ #429800).

-Remove an error from the stap-calling command in TapsetParser
that caused a crash when booting Eclipse with imported tapsets
(EBZ #429597). Also allow the parser to capture a script's error
output, in case the output of a verbose run is ever needed.

-Modify OpenFileAction to allow the Probe & Function views
to open a file without user interaction, and without forcing
the file's editor to use a PathEditorInput (which may open a
duplicate editor).

Change-Id: I0820534a2bc6c1d79ed75570d29d4992b05d920d
Signed-off-by: Andrew Ferrazzutti <aferrazz@redhat.com>
Reviewed-on: https://git.eclipse.org/r/23075
Reviewed-by: Alexander Kurtakov <akurtako@redhat.com>
IP-Clean: Alexander Kurtakov <akurtako@redhat.com>
Tested-by: Alexander Kurtakov <akurtako@redhat.com>