tmf: Fold AlphaNumAttritbute class into the main one
Initially it was planned to allow for different types of attributes,
which could have children whose name could only be numbers, or
sequences of alphanumerical characters.

In the end, it was too complicated to expose this functionality in
StateSystem's API, so to remain as generic as possible, all attributes
are of type "alphanumerical".

The separate class is now unneeded. This also simplifies synchronizing
the container of sub-attributes (which should fix bug #431706).

Change-Id: I626bd07de660131ef6432cb10b0e783201c24782
Signed-off-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>
Reviewed-on: https://git.eclipse.org/r/24275
Tested-by: Hudson CI
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>