tmf: Fix Filter viewer context menu when opened from Colors view
When the filter viewer was opened from the Colors view, the context
menu did not appear anymore. This is because the context menu code
was moved to the FilterView from the FilterViewer so it would only
be executed opening the Filters view. The fix moves the code back
to the viewer and exposes the menu to the view so that it can add
other menu items through extentions.

Change-Id: I0fd447186342bfded6fb202590bbf0e69af6236a
Signed-off-by: Marc-Andre Laperle <marc-andre.laperle@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/25689
Tested-by: Hudson CI