tmf: Fix custom parser wizard preview table
Since the preview table is built using a temporary trace file with a
fixed name, its persistent index was not being rebuilt when the parser
and/or the preview input was changed.

The preview table is now built using a trace class implementation that
is overridden to use the in-memory checkpoint indexer.

Change-Id: I5cd04f26dea7ab005823317fed68130477bf8e0d
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-on: https://git.eclipse.org/r/27124
Reviewed-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
Tested-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
Tested-by: Hudson CI