Systemtap: Improve "Save Log" action.
Improve the UI of the "Save Log" button on script consoles to
better indicate when and to where script output is being logged,
and to allow logs to be saved while scripts are running.

Also, to make these improvements work, fix a bug that prevents
stap script Commands from being disposed when a running script
is re-run.

Change-Id: I38d1d154dba0dcbae5187aa9af8fc31c1672ee5c
Signed-off-by: Andrew Ferrazzutti <aferrazz@redhat.com>
Reviewed-on: https://git.eclipse.org/r/29767
Tested-by: Hudson CI
Reviewed-by: Alexander Kurtakov <akurtako@redhat.com>
Tested-by: Alexander Kurtakov <akurtako@redhat.com>