tmf: Introduce ITmfEventTableColumns and its extension point
Instead of accidentally specifying the whole event table, trace types
can now only define the columns they want. This allows more (most)
trace types to use the default TmfEventsTable, which in turn allows more
trace types to be "aggregated" together in one table when put together
in the same experiment.

The previous "eventsTableType" extension remains there,
 1) for compatibility purposes, but also
 2) for trace types that do very specific things with their event
    table and need to override the default one (ie, GdbTrace).

If a trace type provides both extension points, we'll first try to give
it its event table (ignoring the ITmfEventTableColumns info), and if we
can't because we're within an experiment for instance, then we'll fallback
to just use the columns.

Change-Id: Ie0c7fc9af62dc79aa3add8a605e02455e0df30c4
Signed-off-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>
Reviewed-on: https://git.eclipse.org/r/31407
Tested-by: Hudson CI