TMF: added support for process filter in the TimeGraphFilterDialog.
On top of the filter dialog a text field enables filtering over the
tree of process names.

A node of the tree is shown if:
- the node matches with the search pattern
- one of the children of the node matches with the search pattern
- one of the parents of the node matches with the search pattern

The buttons check-all and check-subtree check only the visible items.
The button uncheck-subtree unchecks all the children (the hidden
children too, since it makes no sense a checked child with an unchecked
parent).
The button uncheck-all unchecks everything, except the hidden roots
(and corresponding subtrees).

Change-Id: I517f3333bb7b6da9fd14eaaac90e7b914671479e
Signed-off-by: Generoso Pagano <generoso.pagano@gmail.com>
Reviewed-on: https://git.eclipse.org/r/32404
Tested-by: Hudson CI
Reviewed-by: Patrick Tasse <patrick.tasse@gmail.com>
Tested-by: Patrick Tasse <patrick.tasse@gmail.com>