rcp: Make archive file name compatible with Maven 3.2.2
The build timestamp format changed in Maven 3.2.2. See
https://jira.codehaus.org/browse/MNG-5452

This means the archive name now looks like
linuxtools-trace-viewer-1.0.0-2014-09-05T21:50:43Z.linux.gtk.x86.tar.gz

This is especially problematic on Windows because colons are used to
specify devices. The format can be changed by to the original format with
a property.

Change-Id: I6c0ec7cc0e8a18ef6a20fb43cc19eadbed80670f
Signed-off-by: Marc-Andre Laperle <marc-andre.laperle@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/32961
Tested-by: Hudson CI