lttng: More performance improvements to the Resources view
In the Resources view, the CPU Status states are further split, when the
current thread switches mid-state, into sub-states by a drawing overlay.

There was no limit to the number of sub-states that this overlay
produces. It should be limited to at most one sub-state per pixel.

Furthermore the vertical bar delimiting these sub-states should not be
drawn when the sub-state is smaller than a pixel, otherwise the bars
can block the underlying state color.

Change-Id: I26fa3ef27ebbde41481fb210c702ddbf22a10773
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-on: https://git.eclipse.org/r/34883
Tested-by: Hudson CI
Reviewed-by: Marc-Andre Laperle <marc-andre.laperle@ericsson.com>
Tested-by: Marc-Andre Laperle <marc-andre.laperle@ericsson.com>