Gcov: fix failing tests.
Fix the failure in GcovTestCLibrary that was caused by retrieval of
the wrong binary to check coverage of. Resolve this by adding a
specified binary path as a parameter of the programmatic version of
OpenGCAction. (Note that the original SWTBot test used the proper
binary by explicitly entering its path in an OpenGCDialog.)

Allow the Gcov tests to once again annotate C editors with coverage
information, as this was erroneously disabled in a previous patch
to these tests. Also fix a NPE in GcovAnnotationModel that this
change exposed.

Change-Id: Iba82ad510c56ea55391a4241c2d1de2d711aca4d
Signed-off-by: Andrew Ferrazzutti <aferrazz@redhat.com>
Reviewed-on: https://git.eclipse.org/r/47507
Tested-by: Hudson CI
Reviewed-by: Alexander Kurtakov <akurtako@redhat.com>