Bug 489781: Fix enablement for dependent toolbar items.
Docker containers view toolbar items are dependent on each other as well
as the state of containers in order to have a proper enablement state.
It's difficult to retrigger property testers to run based on button
activation so move the logic for these actions out of the extension
point registry and into code.

Change-Id: I39654002fda2714f9dc1c1903a5c8891a5daf085
Reviewed-on: https://git.eclipse.org/r/68571
Tested-by: Hudson CI
Reviewed-by: Jeff Johnston <jjohnstn@redhat.com>