Bug 476387 - Allow for connection settings edition
Added new EditConnection wizard , page and model to
allow for changes of name and connection settings
with the same validation rules as in the NewConnection
wizard.
Changes in a connection name are progragated to all
related launch configurations (run and build images).

Refactored the code to hold the connection settings,
using implementations of IDockerConnectionSettings for
Unix Socket and TCP connections.

Removed equals/hashcode methods in DockerConnection
because we need to be able to look-up a connection
as the key in a map even after the name changed
(otherwise, the Map#containsKey(Object) would returns false)

Added listener interface in Containers and Images views
to update the title when the connection changes.

Added SWTBot tests to verify the behaviour of the
new wizard
Change-Id: Idfa1c82fdecdbcbe8a471a35c544aa813efd3d58
Signed-off-by: Xavier Coulon <xcoulon@redhat.com>
Reviewed-on: https://git.eclipse.org/r/73256
Tested-by: Hudson CI