Bug 495293: Fix IOOBE occuring due to empty list of raw image repoTags.
Even if rawImage.repoTags() is not null, it may still be empty,
triggering the IOOBE later on at repoTags.get(0).

Images with no repoTags were showing up in views because only images
with repoTags of ["<none>:<none>"] were being recognized as untagged.
Fix this by moving boolean definitions for taggedImage,
intermediateImage and danglingImage so that they may be evaluated after
we've adjusted repoTags.

Change-Id: Iec5f8dfcf3252bffdcb0eec7789dbf4a5950b3d7
Reviewed-on: https://git.eclipse.org/r/74399
Reviewed-by: Fred Bricon <fbricon@gmail.com>
Tested-by: Hudson CI
Reviewed-by: Roland Grunberg <rgrunber@redhat.com>