Bug 497460: When no active connection exists, disable Pull and Build.
When no Docker daemon connection is active, the Pull Image and Build
Image handlers should be disabled. This prevents users from clicking on
them and increases likelihood of going to create/select an active
connection. For the Build Image launch shortcut, ask the user if they
wish to bring up the new connection wizard.

Change-Id: I6ec8de1dc98b8ad78b4a829b5a767945163700e7
Reviewed-on: https://git.eclipse.org/r/77638
Tested-by: Hudson CI
Reviewed-by: Roland Grunberg <rgrunber@redhat.com>