Bug 499750 - Add commands in Image Hierarchy view
Added commands on containers in Hierarchy view
Introduced 2 new sub-interfaces to IDockerImageHierarchyNode
to distinguish between IDockerImages and IDockerContainers

NOTE: as a side effect to adding commands to the containers, some
command (STOP, START) that would work on multiple containers now
require that all containers be in a valid state to be enabled.

Simplified the code of utility methods in LabelProviderUtils class
that provide StyledString

Splitted clipboard tests in NewDockerConnectionSWTBotTest to test
separately DND.CLIPBOARD and DND.SELECTION_CLIPBOARD. The later
only seems to be a valid case on Linux, so the test is disabled on
other platforms.

Change-Id: I3da20e4184918b09b73a5808b84cd541b5f7bdeb
Signed-off-by: Xavier Coulon <xcoulon@redhat.com>
Reviewed-on: https://git.eclipse.org/r/79650
Tested-by: Hudson CI
Reviewed-by: Jeff Johnston <jjohnstn@redhat.com>