Bug 560077 - Deprecate and remove dataviewers.chart and dataviewers.piechart bundles
Switch Valgrind Massif chart to InteractiveChart. Adds the additional
capability of zoom in/out and other configurations to the chart.

This moves the save button from the view directly to the chart. It's UI
change but considered the added functionality and the ability to export
just zoomed part makes it worth.

Change-Id: Icc3bd4717f984bac7eed3f2282c9af939adb6dea
Signed-off-by: Alexander Kurtakov <akurtako@redhat.com>
Reviewed-on: https://git.eclipse.org/r/157618
Tested-by: Linux Tools Bot <linuxtools-bot@eclipse.org>